feat(sidebar): refactor AddProjectDialog into modular tabs
- Extract GitHub, Local, and QuickStart tabs into separate components
- Add git repository initialization with `git init` and empty initial commit
- Add file system path existence check utility
- Support creating new projects from QuickStart tab with validation
