+++ src/renderer/styles/base.css
@@ -78,6 +78,22 @@ body.resizing .right-panel {
   color: var(--text-secondary);
 }
 
+.sidebar-add-btn {
+  font-size: var(--text-base);
+  color: var(--text-secondary);
+  background: none;
+  border: none;
+  cursor: pointer;
+  padding: var(--space-2) var(--space-6);
+  border-radius: var(--radius-sm);
+  transition: color var(--duration-fast), background var(--duration-fast);
+}
+
+.sidebar-add-btn:hover {
+  color: var(--text-primary);
+  background: var(--bg-hover);
+}
+
 .sidebar-content {
   flex: 1;
   overflow-y: auto;
