+++ src/renderer/styles/index.css
@@ -58,3 +58,4 @@
 @import './analytics.css';
 @import './onboarding.css';
 @import './spotlight-tour.css';
+@import './quick-start.css';
