+++ src/renderer/styles/quick-start.css
@@ -1,0 +1,70 @@
+/* ── Add Project dialog overrides ─────────────────────────────────────────── */
+
+.dialog--add-project {
+  min-width: 520px;
+  max-width: 600px;
+}
+
+.dialog--add-project .dialog-tab {
+  font-size: var(--text-lg);
+  padding: var(--space-8) var(--space-14);
+}
+
+.dialog--add-project .dialog-field label {
+  font-size: var(--text-lg);
+}
+
+.dialog--add-project .dialog-field input {
+  font-size: var(--text-xl);
+}
+
+.dialog--add-project .dialog-tabs + .dialog-field {
+  margin-top: var(--space-8);
+}
+
+/* ── Quick Start template cards ──────────────────────────────────────────── */
+
+.template-grid {
+  display: flex;
+  gap: var(--space-8);
+  margin-top: var(--space-8);
+}
+
+.template-card {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: var(--space-4);
+  padding: var(--space-12) var(--space-16);
+  border: 1px solid var(--border);
+  border-radius: var(--radius);
+  background: var(--bg-secondary);
+  cursor: pointer;
+  transition: border-color var(--duration-fast), background var(--duration-fast);
+  min-width: 80px;
+}
+
+.template-card:hover:not(:disabled) {
+  border-color: var(--text-tertiary);
+}
+
+.template-card--selected {
+  border-color: var(--accent);
+  background: var(--accent-tint-6);
+}
+
+.template-card:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.template-card__icon {
+  font-size: var(--text-2xl);
+  line-height: 1;
+}
+
+.template-card__name {
+  font-size: var(--text-base);
+  color: var(--text-primary);
+  font-weight: var(--weight-medium);
+}
